backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
google-auth==2.35.0
requests==2.32.3
pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.27.2
pytest==8.3.3
